Change bundle interface to use RPC
Summary:
The current bundle interface relies on opening a new python process for each bundle call and parsing parameters to the bundle via SHM and then parsing the stdout to get the result. Needless to say this is rubbish for a variety of reasons - the biggest of which is that the performance is miserable. On system's which have slow IOPS for small files, loading a large interpreter and bundle can be really slow, compounded by being reloaded every time a bundle call is made. Gross.

This new implementation is significantly better. It relies on the xmlrpc module and unix sockets. The rough process is that when a bundle call is first made, the code checks to see if the rpc server is running or not, and if not starts it. If it's already running the existing RPC socket is used. This means the bundle no longer needs the python interpreter to be launched for each bundle call. Reload of changed bundle files is still implemented, but is limited only to the bundle files, and not to the venv.

Test Plan: Tests have been added which test the set up of the RPC server and it's function. Reloading is also tested

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D680