Reduce unix socket path length for bundle communication
Summary: There was a bug where the path to the unix socket for communicating with the job controller could be too long to be valid. This diff changes the logic of how the path to the socket is generated by placing the socket in the temporary directory (/tmp on *nix/BSD), and uses a hash of the parameters to generate the unique socket file name. It then protects the socket using user R/W only on the socket.

Test Plan: Tests still pass that test the socket server logic, and an additional assert has been placed in the code to prevent corner cases triggering unexpected issues.

Reviewers: mr-asher, TomReichardt

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D736